crud and auth added

+++ controllers/contactControllers.js
@@ -1,16 +1,18 @@
 const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel");
 
-//desc Create New Contact
+//desc get all Contacts
 //route GET /api/contacts/
-//access public
+//access private
 
 const getAllContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({user_id: req.user.id});
   res.status(200).json(contacts);
 });
 
-
+//desc Create New Contact
+//route POST /api/contacts/
+//access private
 const createContact = asyncHandler(async (req, res) => {
     const { name, email, phone } = req.body;
     if(!name || !email || !phone) {
@@ -21,11 +23,14 @@ const createContact = asyncHandler(async (req, res) => {
       name,
       email,
       phone,
+      user_id: req.user.id
     })
   res.status(201).json(contact);
 });
 
-
+//desc get single Contact data
+//route GET /api/contacts/:id
+//access private
 const getContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if(!contact){
@@ -35,12 +40,18 @@ const getContact = asyncHandler(async (req, res) => {
   res.status(200).json(contact);
 });
 
-
+//desc update single contact
+//route PUT /api/contacts/:id
+//access private
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if(!contact){
     res.status(404);
     throw new Error("contact not found");
+  }
+  if(contact.user_id !== req.user.id){
+    res.send(403);
+    throw new Error("You don't have permission to update other's contacts")
   }
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
@@ -51,12 +62,18 @@ const updateContact = asyncHandler(async (req, res) => {
 });
 
 
-
+//desc delete single contact
+//route GET /api/contacts/:id
+//access private
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if(!contact){
     res.status(404);
     throw new Error("contact not found");
+  }
+  if(contact.user_id !== req.user.id){
+    res.send(403);
+    throw new Error("You don't have permission to delete other's contacts")
   }
   await Contact.deleteOne({ "_id" : req.params.id });
   res.status(200).json(contact);
